// src/components/MiniPlayer/types.ts
export interface PlayerSize {
	/** player width in px */
	width?: string
	/** player height in px */
	height?: string
}

export interface PlayerPosition {
	/** player top position in px */
	top?: string
	/** player right position in px */
	right?: string
	/** player bottom position in px */
	bottom?: string
	/** player left position in px */
	left?: string
}